+++ sdk/cosmos/azure-cosmos-kafka-connect/src/test/java/com/azure/cosmos/kafka/connect/implementation/source/MetadataMonitorThreadTest.java
@@ -56,6 +56,7 @@ public class MetadataMonitorThreadTest extends KafkaCosmosTestSuiteBase {
                 true,
                 new ArrayList<String>(),
                 new ArrayList<String>());
+
         CosmosMetadataConfig metadataConfig =
             new CosmosMetadataConfig(500, "_cosmos.metadata.topic");
         SourceConnectorContext sourceConnectorContext = Mockito.mock(SourceConnectorContext.class);